Keeping not JSON serializable objects out of html context (#47)
ABlog objects are not JSON serializable. They are used for
creating archive pages, and work fine with HTML builders.
When json builder is used, it pickles HTML context along
with HTML, and this causes serialization issues. So ABlog
objects (functions, Catalog, Blog instance, etc.) are avoided
when builder is not one of html/dirhtml.