Repair update directive
Somewhere along the line, the update directive broke. The last fix led
to update nodes not being properly checked and confusion as to whether
the first paragraph was supposed to be a title or a paragraph.

This fix goes back and returns everything to the documented update
directive behavior, complete with date checking. The update directive
should be easier to update from this.