fix(feeds): Improve tag/category term format
Remove spaces from the Atom category term values.  I figure since the label has the
full, original ABlog tag value there's no reason not to take advantage of the term to
remove spaces which seems like a version of the tag that is likely to be useful in many
external services.